requirements.txt
APIResponse==0.2
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-localflavor==3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Flask==2.0.1
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
pyparsing==2.4.7
python-stdnum==1.16
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
validate-docbr==1.8.2
Werkzeug==2.0.1
